requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
boto3==1.37.23
botocore==1.37.38
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
croniter==6.0.0
distro==1.9.0
fastapi==0.115.8
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.31.0
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jmespath==1.0.1
mem0ai==0.1.116
multidict==6.6.4
numpy==2.3.2
openai==1.100.1
phonenumbers==9.0.12
portalocker==3.2.0
posthog==6.6.0
propcache==0.3.2
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
qdrant-client==1.15.1
redis==6.4.0
requests==2.32.4
celery==5.4.0
rsa==4.9.1
s3transfer==0.11.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.45.3
tenacity==9.1.2
tqdm==4.67.1
twilio==9.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.20.1
